Rename : global.common -> domain.common으로 위치 변경
