+++ src/main/java/com/kkiri_trip/back/domain/common/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.kkiri_trip.back.global.common.entity;
+package com.kkiri_trip.back.domain.common.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
