+++ src/main/java/com/kkiri_trip/back/domain/user/entity/User.java
@@ -1,6 +1,6 @@
 package com.kkiri_trip.back.domain.user.entity;
 
-import com.kkiri_trip.back.global.common.entity.BaseEntity;
+import com.kkiri_trip.back.domain.common.entity.BaseEntity;
 import com.kkiri_trip.back.global.enums.Gender;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
